feat: refatora repositorio produto
