--- src/repositorio/produto/interfaces/compostas/ProdutoAtivo.java
@@ -1,9 +1,0 @@
-package repositorio.produto.interfaces.compostas;
-
-import repositorio.produto.interfaces.especificas.ProdutoConsulta;
-import repositorio.produto.interfaces.especificas.ProdutoManipulacao;
-import repositorio.produto.interfaces.especificas.ProdutoAlteracao;
-
-public interface ProdutoAtivo extends ProdutoConsulta, ProdutoManipulacao, ProdutoAlteracao {
-
-}
