+++ src/repositorio/produto/interfaces/compostas/RepositorioProduto.java
@@ -3,5 +3,6 @@ package repositorio.produto.interfaces.compostas;
 import repositorio.produto.interfaces.especificas.ProdutoConsulta;
 import repositorio.produto.interfaces.especificas.ProdutoManipulacao;
 
-public interface ProdutoInativo extends ProdutoConsulta, ProdutoManipulacao {
+public interface RepositorioProduto extends ProdutoConsulta, ProdutoManipulacao {
+
 }
