--- src/repositorio/produto/interfaces/especificas/ProdutoAlteracao.java
@@ -1,7 +1,0 @@
-package repositorio.produto.interfaces.especificas;
-
-import dominio.Produto.Produto;
-
-public interface ProdutoAlteracao {
-    void alterarNome(Produto produto, String novoNome);
-}
