+++ src/repositorio/produto/interfaces/especificas/ProdutoConsulta.java
@@ -5,7 +5,9 @@ import dominio.Produto.Produto;
 import java.util.List;
 
 public interface ProdutoConsulta {
-    List<Produto> listarProdutos();
+    List<Produto> listarProdutosInativos();
+    List<Produto> listarProdutosAtivos();
+    List<Produto> listarTodosProdutos();
     Produto buscarPorNome(String nome);
     Produto buscarPorId(long id);
 }
